backend/src/requirements.txt
Flask
mysql-connector-python 